working on separating sounds (2D & 3D)

+++ src/model_entity_color_renderer.cpp
@@ -96,7 +96,7 @@ void ModelEntityColorRenderer::processPointlightEntities(const unordered_map<str
 	}
 
 	// Upload lights
-	for(unsigned int i = 0; i < visibleEntities.size(); i++)
+	for(size_t i = 0; i < visibleEntities.size(); i++)
 	{
 		_shader.uploadUniform("u_pointlightPositions[" + to_string(i) + "]", visibleEntities[i]->getPosition());
 		_shader.uploadUniform("u_pointlightColors[" + to_string(i) + "]", visibleEntities[i]->getColor());
@@ -122,7 +122,7 @@ void ModelEntityColorRenderer::processSpotlightEntities(const unordered_map<stri
 	}
 
 	// Upload lights
-	for(unsigned int i = 0; i < visibleEntities.size(); i++)
+	for(size_t i = 0; i < visibleEntities.size(); i++)
 	{
 		_shader.uploadUniform("u_spotlightPositions[" + to_string(i) + "]", visibleEntities[i]->getPosition());
 		_shader.uploadUniform("u_spotlightFrontVectors[" + to_string(i) + "]", visibleEntities[i]->getFrontVector());
